feat(entry): Introduce optional metadata to Entries
Adds an optional `metadata` field to `Entry` objects. This metadata is
not part of the main data model and is not merged between entries.
Its primary purpose is to store information that can improve the
efficiency of certain operations and for experimentation.

Currently, if an operation does not modify the "settings" subtree,
metadata is added to the new entry containing references (tips) of the
current "settings" subtree. This allows for more efficient verification
of settings in sparse checkout scenarios without requiring traversal
of the entire history graph.